+++ NEWS.md
@@ -1,7 +1,7 @@
 # h5lite 2.0.0.4
 
-* Added support for SZIP compression.
-* Added support for partial reading.
+* Added support for SZIP compression - see `vignette('data-compression')`.
+* Added support for partial reading - see `vignette('partial-reading')`.
 * `h5_length()` and `h5_dim()` now return numeric values instead of integer.